+++ tests/test_compression.cpp
@@ -80,7 +80,7 @@ static cz::Input_File run_compression_script(char compressed_file_buffer[L_tmpna
 template <class DecompressionStream>
 static void test_decompression(const char* script) {
     size_t length = GENERATE(0, 10, -1);
-    if (length == -1) {
+    if (length == (size_t)-1) {
         std::mt19937 gen(dev());
         std::uniform_int_distribution<> dist(0, 1 << 20);
         length = dist(gen);
@@ -110,7 +110,7 @@ static void test_decompression(const char* script) {
 template <class CompressionStream>
 static void test_compression(const char* script) {
     size_t length = GENERATE(0, 10, -1);
-    if (length == -1) {
+    if (length == (size_t)-1) {
         std::mt19937 gen(dev());
         std::uniform_int_distribution<> dist(0, 1 << 20);
         length = dist(gen);
